Moves callbacks for events outside main.js
